feat(client): Extract Milestone feed to own React functional component
