+++ client/src/app/routes/stardust/landing/Landing.tsx
@@ -1,9 +1,7 @@
 import classNames from "classnames";
-import moment from "moment";
 import React, { ReactNode } from "react";
 import { Link, RouteComponentProps } from "react-router-dom";
 import { ServiceFactory } from "../../../../factories/serviceFactory";
-import { DateHelper } from "../../../../helpers/dateHelper";
 import { isShimmerNetwork } from "../../../../helpers/networkHelper";
 import { RouteBuilder } from "../../../../helpers/routeBuilder";
 import { INetwork } from "../../../../models/config/INetwork";
@@ -12,15 +10,14 @@ import { STARDUST } from "../../../../models/config/protocolVersion";
 import { IFeedItem } from "../../../../models/feed/IFeedItem";
 import { IFeedItemMetadata } from "../../../../models/feed/IFeedItemMetadata";
 import { NetworkService } from "../../../../services/networkService";
-import FeedMilestoneInfo from "../../../components/FeedMilestoneInfo";
 import Feeds from "../../../components/stardust/Feeds";
-import Tooltip from "../../../components/Tooltip";
 import NetworkContext from "../../../context/NetworkContext";
 import { LandingRouteProps } from "../../LandingRouteProps";
 import AnalyticStats from "./AnalyticStats";
 import FeedFilters from "./FeedFilters";
 import InfoBox from "./InfoBox";
 import { FeedTabs, getDefaultLandingState, LandingState } from "./LandingState";
+import MilestoneFeed from "./MilestoneFeed";
 import "./Landing.scss";
 
 /**
@@ -108,85 +105,16 @@ class Landing extends Feeds<RouteComponentProps<LandingRouteProps>, LandingState
                                         ))}
                                     </div>
                                     {currentTab === FeedTabs.MILESTONES && (
-                                        <>
-                                            <div className="section--header row space-between padding-l-8">
-                                                <h2>Latest milestones</h2>
-                                            </div>
-                                            <div className="feed-items">
-                                                <div className="row feed-item--header ms-feed">
-                                                    <span className="label ms-index">Index</span>
-                                                    <span className="label ms-id">Milestone Id</span>
-                                                    <span className="label ms-blocks">Blocks</span>
-                                                    <span className="label ms-txs">Txs</span>
-                                                    <span className="label ms-timestamp">Timestamp</span>
-                                                </div>
-                                                {milestones.map(ms => {
-                                                    const index = ms.properties?.index as number;
-                                                    const milestoneId = ms.properties?.milestoneId as string;
-                                                    const milestoneIdShort =
-                                                        `${milestoneId.slice(0, 6)}....${milestoneId.slice(-6)}`;
-                                                    const timestamp = ms.properties?.timestamp as number * 1000;
-                                                    const includedBlocks = 5;
-                                                    const txs = 2;
-                                                    const ago = moment(timestamp).fromNow();
-                                                    const tooltipContent = DateHelper.formatShort(timestamp);
-
-                                                    return (
-                                                        <div className="feed-item ms-feed" key={ms.id}>
-                                                            <div className="feed-item__content">
-                                                                <span className="feed-item--label">Index</span>
-                                                                <span className="feed-item--value ms-index">
-                                                                    <Link
-                                                                        className="feed-item--hash ms-id"
-                                                                        to={`${network}/search/${index}`}
-                                                                    >
-                                                                        {index}
-                                                                    </Link>
-                                                                </span>
-                                                            </div>
-                                                            <div className="feed-item__content">
-                                                                <span className="feed-item--label">Milestone id</span>
-                                                                <Link
-                                                                    className="feed-item--hash ms-id"
-                                                                    to={RouteBuilder.buildItem(
-                                                                        networkConfig, milestoneId
-                                                                    )}
-                                                                >
-                                                                    {milestoneIdShort}
-                                                                </Link>
-                                                            </div>
-                                                            <div className="feed-item__content">
-                                                                <span className="feed-item--label">Blocks</span>
-                                                                <span className="feed-item--value ms-blocks">
-                                                                    {includedBlocks}
-                                                                </span>
-                                                            </div>
-                                                            <div className="feed-item__content">
-                                                                <span className="feed-item--label">Txs</span>
-                                                                <span className="feed-item--value ms-txs">
-                                                                    {txs}
-                                                                </span>
-                                                            </div>
-                                                            <div className="feed-item__content">
-                                                                <span className="feed-item--label">Timestamp</span>
-                                                                <span className="feed-item--value ms-timestamp">
-                                                                    <Tooltip
-                                                                        tooltipContent={tooltipContent}
-                                                                    >
-                                                                        {ago}
-                                                                    </Tooltip>
-                                                                </span>
-                                                            </div>
-                                                        </div>
-                                                    );
-                                                })}
-                                            </div>
-                                        </>
+                                        <MilestoneFeed
+                                            networkConfig={networkConfig}
+                                            milestones={milestones}
+                                            latestMilestoneIndex={latestMilestoneIndex}
+                                        />
                                     )}
                                     {currentTab === FeedTabs.BLOCKS && (
                                         <>
                                             <div className="section--header row space-between padding-l-8">
-                                                <h2>Latest blocks</h2>
+                                                <h2 />
                                                 <FeedFilters
                                                     networkConfig={networkConfig}
                                                     settingsService={this._settingsService}
@@ -196,10 +124,6 @@ class Landing extends Feeds<RouteComponentProps<LandingRouteProps>, LandingState
                                                     }}
                                                 />
                                             </div>
-                                            <FeedMilestoneInfo
-                                                milestoneIndex={latestMilestoneIndex}
-                                                frequencyTarget={networkConfig.milestoneInterval}
-                                            />
                                             <div className="feed-items">
                                                 <div className="row feed-item--header">
                                                     <span className="label">Block id</span>
